update: 사용자 인증 커맨드 핸들러 위치 이동
