+++ hobbyroom/auth/service/__init__.py
@@ -1,1 +1,2 @@
+from .command_handler import *  # noqa: F401, F403
 from .jwt_handler import *  # noqa: F401, F403
